add user roles navigation and dashboard

+++ app/src/main/java/com/keneth/realestateapplication/views/LoginScreen.kt
@@ -2,10 +2,7 @@ package com.keneth.realestateapplication.views
 
 import androidx.compose.runtime.Composable
 import androidx.navigation.NavController
-import androidx.navigation.NavHostController
-import com.keneth.realestateapplication.viewModels.AuthStatus
-import com.keneth.realestateapplication.viewModels.UserViewModel
-import android.content.Context
+import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
@@ -14,7 +11,7 @@ import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material3.*
 import androidx.compose.runtime.*
-
+import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
@@ -28,11 +25,11 @@ import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.lifecycle.viewmodel.compose.viewModel
 import com.keneth.realestateapplication.R
 import com.keneth.realestateapplication.UserPreferences
+import com.keneth.realestateapplication.viewModels.AuthStatus
+import com.keneth.realestateapplication.viewModels.UserViewModel
 import kotlinx.coroutines.delay
-import kotlinx.coroutines.launch
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -41,63 +38,69 @@ fun LoginScreen(
     viewModel: UserViewModel = viewModel()
 ) {
     val context = LocalContext.current
-    var email by remember { mutableStateOf("") }
-    var password by remember { mutableStateOf("") }
-    var showErrorSnackbar by remember { mutableStateOf(false) }
-    var showSuccessSnackbar by remember { mutableStateOf(false) }
-    var errorMessage by remember { mutableStateOf<String?>(null) }
+    var email by rememberSaveable { mutableStateOf("") }
+    var password by rememberSaveable { mutableStateOf("") }
+    var showMessage by remember { mutableStateOf(false) }
+    var message by remember { mutableStateOf<String?>(null) }
+    var isSuccess by remember { mutableStateOf(false) }
     val authState = viewModel.authState.value
-    val isLoading = viewModel.isLoading.value
-
 
     val profileFontFamily = FontFamily(
         Font(R.font.darkmode_regular_400, weight = FontWeight.Normal),
         Font(R.font.cluisher_brush, weight = FontWeight.Bold)
     )
+
+    // Handling authentication state changes
     LaunchedEffect(authState) {
         when (authState) {
             is AuthStatus.SuccessWithData -> {
                 val token = authState.user?.get("token").toString()
-                println("Token from authState: $token")
-                if (token.isNotEmpty()) {
-                    // Store token
+                val role = authState.user?.get("userRole").toString()
+                if (token.isNotEmpty() && role.isNotEmpty()) {
                     UserPreferences.deleteToken(context)
                     UserPreferences.storeToken(context, token)
-
-                    showSuccessSnackbar = true
-                    delay(2000) // Show success message before navigating
-                    navController.navigate(Screen.Dashboard.route) {
-                        popUpTo(Screen.Login.route) { inclusive = true }
+                    UserPreferences.deleteUserRole(context)
+                    UserPreferences.storeUserRole(context, role)
+                    message = "Login successful!"
+                    isSuccess = true
+                    showMessage = true
+                    delay(2000)
+                    when (role) {
+                        "ADMIN" -> navController.navigate(Screen.Dashboard.route)
+                        "AGENT" -> navController.navigate(Screen.AgentDashboard.route)
+                        "BUYER" -> navController.navigate(Screen.BuyerDashboard.route)
+                        "SELLER" -> navController.navigate(Screen.SellerDashboard.route)
+                        "LANDLORD" -> navController.navigate(Screen.LandLordDashboard.route)
+                        "TENANT" -> navController.navigate(Screen.TenantDashboard.route)
+                        "GUEST" -> navController.navigate(Screen.GuestDashboard.route)
+                        else -> navController.navigate(Screen.Login.route)
                     }
                 } else {
-                    println("Token is empty, not navigating to dashboard")
+                    message = "Token or role is empty"
+                    isSuccess = false
+                    showMessage = true
                 }
             }
 
             is AuthStatus.Error -> {
-                errorMessage = authState.message
-                showErrorSnackbar = true
+                message = authState.message
+                isSuccess = false
+                showMessage = true
             }
 
             else -> {}
         }
     }
+
     Scaffold(
         modifier = Modifier.fillMaxSize(),
-        snackbarHost = {
-            SnackbarHost(hostState = remember { SnackbarHostState() }) { snackbarData ->
-                Snackbar(
-                    snackbarData = snackbarData,
-                    containerColor = if (showErrorSnackbar) Color.Red else Color.Green
-                )
-            }
-        },containerColor = Color.White, contentColor = Color.Black,
+        containerColor = Color.White,
+        contentColor = Color.Black,
     ) { paddingValues ->
         Box(
             modifier = Modifier
                 .fillMaxSize()
                 .padding(paddingValues),
-            //    .background(Color.White),
             contentAlignment = Alignment.Center
         ) {
             Column(
@@ -113,18 +116,16 @@ fun LoginScreen(
                         fontSize = 40.sp,
                         fontFamily = profileFontFamily,
                         fontWeight = FontWeight.Normal,
-                        color = Color.Green, letterSpacing = 5.0.sp
+                        color = Color.Green,
+                        letterSpacing = 5.0.sp
                     )
-
                 )
                 Spacer(modifier = Modifier.height(50.dp))
-//               sp Spacer(modifier = Modifier.height(16.dp))
-//                Image(
-//                    painter = painterResource(id = R.drawable.img10),
-//
-//                    contentDescription = "Splash Image",
-//                    modifier = Modifier.size(250.dp)
-//                )
+                Image(
+                    painter = painterResource(id = R.drawable.img10), // Ensure this drawable exists
+                    contentDescription = "Splash Image",
+                    modifier = Modifier.size(250.dp)
+                )
                 Spacer(modifier = Modifier.height(16.dp))
                 OutlinedTextField(
                     value = email,
@@ -134,9 +135,13 @@ fun LoginScreen(
                     keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Email)
                 )
                 Spacer(modifier = Modifier.height(16.dp))
-                PasswordField(
-                    password = password,
-                    onPasswordChange = { password = it }
+                OutlinedTextField(
+                    value = password,
+                    onValueChange = { password = it },
+                    label = { Text("Password") },
+                    modifier = Modifier.fillMaxWidth(),
+                    keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Password),
+                    visualTransformation = PasswordVisualTransformation()
                 )
                 Spacer(modifier = Modifier.height(16.dp))
                 Button(
@@ -153,37 +158,26 @@ fun LoginScreen(
                         defaultElevation = 6.dp,
                         pressedElevation = 2.dp,
                         hoveredElevation = 8.dp
-                    )
-                ) {
-                    if (isLoading) {
-                        Text("Logging In...")
-                    }
-                    Text("Login")
-                }
-                // Login Button
-                Button(
-                    onClick = { viewModel.login(email, password) },
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(vertical = 8.dp),
-                    shape = RoundedCornerShape(4.dp),
-                    colors = ButtonDefaults.buttonColors(
-                        containerColor = Color(0xFFCB35D5),
-                        contentColor = Color.White
                     ),
-                    elevation = ButtonDefaults.buttonElevation(
-                        defaultElevation = 6.dp,
-                        pressedElevation = 2.dp,
-                        hoveredElevation = 8.dp
-                    )
+                    enabled = authState != AuthStatus.Loading // Disable button during loading
                 ) {
-
-                    Text("Login with Google")
+                    Row(
+                        verticalAlignment = Alignment.CenterVertically,
+                        horizontalArrangement = Arrangement.Center
+                    ) {
+                        if (authState == AuthStatus.Loading) {
+                            CircularProgressIndicator(
+                                color = Color.White,
+                                modifier = Modifier.size(20.dp)
+                            )
+                            Spacer(modifier = Modifier.width(8.dp))
+                            Text("Logging In...")
+                        } else {
+                            Text("Login")
+                        }
+                    }
                 }
-
                 Spacer(modifier = Modifier.height(16.dp))
-
-                // Sign Up Link
                 Text(
                     "Don't have an account? Sign up.",
                     modifier = Modifier.clickable {
@@ -195,31 +189,52 @@ fun LoginScreen(
         }
     }
 
-    // Show Snackbars
-    if (showErrorSnackbar) {
-        LaunchedEffect(showErrorSnackbar) {
-            delay(3000)
-            showErrorSnackbar = false
+    // Show message dialog based on authState
+    if (showMessage) {
+        LaunchedEffect(Unit) {
+            delay(5000)
+            showMessage = false
         }
-        Snackbar(
-            modifier = Modifier.padding(8.dp),
-            containerColor = Color.Red
+        Box(
+            modifier = Modifier
+                .fillMaxSize()
+                .background(Color.Black.copy(alpha = 0.5f)),
+            contentAlignment = Alignment.Center
         ) {
-            Text(errorMessage ?: "Invalid credentials", color = Color.White)
-        }
-    }
-
-    if (showSuccessSnackbar) {
-        LaunchedEffect(showSuccessSnackbar) {
-            delay(2000)
-            showSuccessSnackbar = false
-        }
-        Snackbar(
-            modifier = Modifier.padding(8.dp),
-            containerColor = Color.Green
-        ) {
-            Text("Login successful!", color = Color.White)
+            Card(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(16.dp)
+                    .height(120.dp), // Increased card height
+                shape = RoundedCornerShape(8.dp),
+                colors = CardDefaults.cardColors(
+                    containerColor = if (isSuccess) Color.Green else Color.Red
+                ),
+                elevation = CardDefaults.cardElevation(8.dp)
+            ) {
+                Column(
+                    modifier = Modifier
+                        .fillMaxSize()
+                        .padding(16.dp),
+                    verticalArrangement = Arrangement.Center,
+                    horizontalAlignment = Alignment.CenterHorizontally
+                ) {
+                    Image(
+                        painter = painterResource(
+                            id = if (isSuccess)
+                                R.drawable.correct else R.drawable.error2 // Ensure these drawables exist
+                        ),
+                        contentDescription = if (isSuccess) "Success Icon" else "Failure Icon",
+                        modifier = Modifier.size(50.dp)
+                    )
+                    Spacer(modifier = Modifier.height(8.dp))
+                    Text(
+                        text = message ?: "",
+                        color = Color.White,
+                        style = MaterialTheme.typography.bodyMedium
+                    )
+                }
+            }
         }
     }
 }
-
